Rename InputService to RemoteService and create new InputService
